+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All Notable changes to `gigya-client` will be documented in this file
 
+## 1.0 - 2016-11-28
+
+### Changed
+- Upgrade Guzzle from v5 to v6
+- All Subscribers are now Guzzle Middleware
+- Renamed `->addSubscriber` to `->addHandler`
+- Renamed `->removeSubscriber` to `->removeHandler`
+
+## 0.3 - 2016-04-02
+
+### Added
+- OAuth2 authentication
+
 ## 0.2 - 2015-10-20
 
 ### Changed